generator/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "langtest-gen"
version = "0.1.0"
description = "LLM-powered test generator for LangTest component library"
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "LangTest" }]
dependencies = [
  "langchain-core>=0.2.0",
  "langchain-openai>=0.1.0",
  "pydantic>=2.7.0",
  "pydantic-settings>=2.3.0",
  "click>=8.1.0",
  "rich>=13.7.0",
  "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2.0",
  "pytest-asyncio>=0.23.0",
]

[project.scripts]
langtest-gen = "langtest_gen.cli:main"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
